Reading data from 2 encoders

+++ encoder_test/Core/Inc/encoder.h
@@ -16,7 +16,7 @@ typedef struct {
 } Encoder;
 
 void ENCODER_init(Encoder* encoder, TIM_HandleTypeDef* htim);
-uint32_t ENCODER_get_value(Encoder* encoder);
-uint32_t ENCODER_get_change(Encoder* encoder);
+uint16_t ENCODER_get_value(Encoder* encoder);
+int8_t ENCODER_get_change(Encoder* encoder);
 
 #endif /* INC_ENCODER_H_ */
